backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cloudinary>=1.41.0
slowapi>=0.1.9
python-multipart>=0.0.12
python-dotenv>=1.0.1
